Modify q_insert_head and q_insert_tail
Insert node at head/tail of a queue

The list_add and list_add_tail APIs in list.h use two pointer of
list_head struct. One is a pointer to the new node 'node' and
the other is a pointer to the head of the list where the insertion
is to take place.

These two APIs help us to insert element_t struct to the head/tail
of a queue.