Modify swap()
Swap every two adjacent nodes

The node to be swapped is pointed to by list_head struct. Utllize
list_move_tail to achieve this objective, which is composed of
list_del for removing the node from the list, and list_add_tail
for adding the node to the tail of the circular list.