Modify reverseK function
Reverse the nodes of the list k at a time

If k is less than or equal zero, do nothing. And if k is large than
the length of the list, do nothing.

Here use the list_cut_position to unlink the list into the sub-lists
with k length then reverse each sub-lists, finally use
list_splice_tail_init to splice each sub-list to one list.