Implement basic queue functions
The function 'q_new' is designed to create an empty queue
using 'malloc'. Since the allocated memory is uninitialized and
its availability is not guaranteed[1], the macro 'INIT_LIST_HEAD' is
used to initialize the next and prev pointers to point to list_head
itself, ensuring that the allocated memory is properly initialized.

The function 'q_insert_head' is designed to insert an element to the
beginning of the queue to be inserted. To prevent 's' from being
modified by other processes, a separate block of memory needs to
be allocated, and a copy of 's' should be made as an independent string
for operations. This copied string is then used for insertion.
Here, 'element_t' is utilized to implement this process, as the
element_t includes a 'list_head'. The advantage of this approach is that
by operating only on 'list_head', the API provided in list.h can
be used to perform operations on a circular linked list, including
list_add, which directly inserts the element at the beginning of
the queue.

The function 'q_free' is designed to free all memory in the queue.
Releasing the memory used by the queue requires freeing two parts:
(1) The memory occupied by the queue's own list_head.
(2) The memory occupied by each element in the queue.
Since a circular linked list can lead to a use-after-free issue
if memory is freed without proper handling, list.h provides the
'list_for_each_entry_safe' API to safely traverse the elements in
the queue. This API uses the *safe pointer to keep track of the
next entry, allowing elements to be safely removed during the iteration
process.

Refs:
    1. 'malloc' — Linux manual page

Change-Id: I64940c3a983597eb37d44b17085759dca21b1b4d